requirements.txt
opencv-python
pillow
numpy
pytesseract 
fastapi
uvicorn